requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
overpy>=0.6
cachetools>=5.3.0
typing-extensions>=4.8.0
